updated api/longId route
