+++ server/controllers/assets/utils/getClaimIdAndServeAsset.js
@@ -1,13 +1,17 @@
+const logger = require('winston');
+
 const db = require('../../../models');
 
 const getClaimId = require('../../api/claim/longId/getClaimId.js');
 const { handleErrorResponse } = require('../../utils/errorHandlers.js');
 
-const serveAssetToClient = require('./serveAssetToClient.js');
+const getLocalFileRecord = require('./getLocalFileRecord.js');
+const serveFile = require('./serveFile.js');
 
 const NO_CHANNEL = 'NO_CHANNEL';
 const NO_CLAIM = 'NO_CLAIM';
 const BLOCKED_CLAIM = 'BLOCKED_CLAIM';
+const NO_FILE = 'NO_FILE';
 
 const getClaimIdAndServeAsset = (channelName, channelClaimId, claimName, claimId, originalUrl, ip, res) => {
   getClaimId(channelName, channelClaimId, claimName, claimId)
@@ -16,26 +20,36 @@ const getClaimIdAndServeAsset = (channelName, channelClaimId, claimName, claimId
       return db.Blocked.isNotBlocked(fullClaimId, claimName);
     })
     .then(() => {
-      serveAssetToClient(claimId, claimName, res);
+      return getLocalFileRecord(claimId, claimName);
+    })
+    .then(fileRecord => {
+      serveFile(fileRecord, res);
     })
     .catch(error => {
       if (error === NO_CLAIM) {
+        logger.debug('no claim found');
         return res.status(404).json({
           success: false,
-          message: 'No claim id could be found',
+          message: 'No matching claim id could be found for that url',
         });
       }
       if (error === NO_CHANNEL) {
+        logger.debug('no channel found');
         return res.status(404).json({
           success: false,
-          message: 'No channel id could be found',
+          message: 'No matching channel id could be found for that url',
         });
       }
       if (error === BLOCKED_CLAIM) {
+        logger.debug('claim was blocked');
         return res.status(410).json({
           success: false,
           message: 'In response to a complaint we received under the US Digital Millennium Copyright Act, we have blocked access to this content from our applications. For more details, see https://lbry.io/faq/dmca',
         });
+      }
+      if (error === NO_FILE) {
+        logger.debug('claim was blocked');
+        return res.status(307).redirect(`/api/claim/get/${name}/${claimId}`);
       }
       handleErrorResponse(originalUrl, ip, error, res);
     });
