+++ server/controllers/assets/utils/serveFile.js
@@ -1,0 +1,14 @@
+const logger = require('winston');
+
+const serveFile = ({ filePath, fileType }, res) => {
+  logger.verbose(`serving file: ${filePath}`);
+  const sendFileOptions = {
+    headers: {
+      'X-Content-Type-Options': 'nosniff',
+      'Content-Type'          : fileType || 'image/jpeg',
+    },
+  };
+  res.status(200).sendFile(filePath, sendFileOptions);
+};
+
+module.exports = serveFile;
